requirements.txt
black==23.3.0
pylint==2.17.4
pytest-cov==4.1.0
requests==2.31.0
